mvc 패턴에 맞춰 코드 수정
