+++ src/main/java/racingcar/Application.java
@@ -1,11 +1,10 @@
 package racingcar;
 
-import racingcar.view.Racing;
+import racingcar.view.RacingView;
 
-public class Application extends Racing {
+public class Application{
 
     public static void main(String[] args) {
-        Racing racing = new Racing();
-        racing.inputNames();
+        RacingView racingView = new RacingView();
     }
 }
